Got much of RangeServer recovery working
Split commit log into three logs:  root, metadata, and user.  Added
crash test framework; mid-split recovery still needs to be implemented and
tested

+++ src/cc/Hypertable/Lib/MetaLogDfsBase.cc
@@ -33,7 +33,7 @@ using namespace Serialization;
 namespace {
   const uint32_t DFS_BUFFER_SIZE = -1;  // use dfs defaults
   const uint32_t DFS_NUM_REPLICAS = -1; // ditto
-  const uint32_t DFS_BLOCK_SIZE = -1;   // ditto
+  const int64_t DFS_BLOCK_SIZE = -1;   // ditto
   const size_t ML_ENTRY_BUFFER_RESERVE = ML_ENTRY_HEADER_SIZE + 512;
 } // local namespace
